WIP: gym simulator dev
