+++ src/simulator/utils.py
